Factor out platform-specific memory management
mmap-based versions -> unix.c; VirtualAlloc-based version -> win32.c
