+++ runtime/caml/osdeps.h
@@ -103,6 +103,13 @@ extern char_os *caml_secure_getenv(char_os const *var);
    cannot be determined, return -1. */
 extern int caml_num_rows_fd(int fd);
 
+/* Memory management platform-specific operations */
+
+void *caml_plat_mem_map(uintnat, uintnat, int);
+void *caml_plat_mem_commit(void *, uintnat);
+void caml_plat_mem_decommit(void *, uintnat);
+void caml_plat_mem_unmap(void *, uintnat);
+
 #ifdef _WIN32
 
 extern int caml_win32_rename(const wchar_t *, const wchar_t *);
