+++ runtime/caml/platform.h
@@ -131,8 +131,9 @@ void caml_plat_cond_free(caml_plat_cond*);
 
 uintnat caml_mem_round_up_pages(uintnat size);
 /* The size given to caml_mem_map and caml_mem_commit must be a multiple of
-   caml_sys_pagesize. The size given to caml_mem_unmap and caml_mem_decommit
-   must match the size given to caml_mem_map/caml_mem_commit for mem. */
+   caml_plat_mmap_granularity. The size given to caml_mem_unmap and
+   caml_mem_decommit must match the size given to caml_mem_map/caml_mem_commit
+   for mem. */
 void* caml_mem_map(uintnat size, uintnat alignment, int reserve_only);
 void* caml_mem_commit(void* mem, uintnat size);
 void caml_mem_decommit(void* mem, uintnat size);
@@ -183,7 +184,7 @@ Caml_inline void caml_plat_unlock(caml_plat_mutex* m)
 
 /* On Windows, the SYSTEM_INFO.dwPageSize is a DWORD (32-bit), but conveniently
    long is also 32-bit */
-extern intnat caml_sys_pagesize;
+extern intnat caml_plat_mmap_granularity;
 
 #endif /* CAML_INTERNALS */
