+++ runtime/win32.c
@@ -1100,7 +1100,7 @@ void caml_init_os_params(void)
      page size and granularity. This simplifies caml_mem_map, since we can
      guarantee that trimming will not be required. */
   CAMLassert(si.dwAllocationGranularity >= si.dwPageSize);
-  caml_sys_pagesize = si.dwAllocationGranularity;
+  caml_plat_mmap_granularity = si.dwAllocationGranularity;
 
   /* Get the number of nanoseconds for each tick in QueryPerformanceCounter */
   QueryPerformanceFrequency(&frequency);
@@ -1113,4 +1113,33 @@ int64_t caml_time_counter(void)
 
   QueryPerformanceCounter(&now);
   return (int64_t)(now.QuadPart * clock_period);
+}
+
+void *caml_plat_mem_map(uintnat size, uintnat alignment, int reserve_only)
+{
+  /* VirtualAlloc returns an address aligned to caml_plat_mmap_granularity, so
+     trimming will not be required. VirtualAlloc returns 0 on error. */
+  if (alignment > caml_plat_mmap_granularity)
+    caml_fatal_error("Cannot align memory to %" ARCH_INTNAT_PRINTF_FORMAT "x"
+                     " on this platform", alignment);
+  return
+    VirtualAlloc(NULL, size,
+                 MEM_RESERVE | (reserve_only ? 0 : MEM_COMMIT),
+                 reserve_only ? PAGE_NOACCESS : PAGE_READWRITE);
+}
+
+void* caml_plat_mem_commit(void* mem, uintnat size)
+{
+  return VirtualAlloc(mem, size, MEM_COMMIT, PAGE_READWRITE);
+}
+
+void caml_plat_mem_decommit(void* mem, uintnat size)
+{
+  VirtualFree(mem, size, MEM_DECOMMIT);
+}
+
+void caml_plat_mem_unmap(void* mem, uintnat size)
+{
+  if (!VirtualFree(mem, 0, MEM_RELEASE))
+    CAMLassert(0);
 }
